SDK Updater: Store local source properties when installing.
When installing an archive, a "source.properties" file is saved
in the directory that contains all the information from the
source (Source URL, package attributes, archive attributes.)

When loading local packages, these properties are used to
recreate the local package attributes if present. This is
also used to indentify local extra packages.