Add support for opening HPROF files in Eclipse.
The action taken on successful HPROF dump can now be configured from
the plug-in preference to be either "save to disk" or "open in eclipse".

If a plug-in supporting .hprof files (such as Eclipse Memory Analyzer Tool)
is present then the file will just be opened in the tool.

"save to disk" does not convert the file, "open in eclipse" does convert
the hprof file (using hprof-conv present in the SDK)

Change-Id: Iceb347abb2af5a2979b6028c22dcbdc3bc7250ca