XML Schema for layout configurations.
The only changes to the original XML sample:
- the root element is called "layout-configs" (plural for container).

- the <default> and <config> elements must be defined in that order:
 - 0..1 of <default>
   folowed by
 - 1..n of <config>
It's invalid for the <device> node to be empty.

default/config are defined by a <xsd:sequence>, which imposes the
strict ordering. At first I used an <xsd:choice> which does not
impose any ordering but then it becomes possible to define an
empty <device> element. I'd rather live with the strict ordering
(which also makes sense anyway) and enforce non-empty <device>
elements in the schema.

BUG 2138329

Change-Id: I581afb1d72825ae79df00d81c74c80a2a18680ad