Emulator launch options dialog.
This is displayed when clicking the Start button in the AVD manager.

The dialog allows to easily scale the emulator to match an arbitrary screen size
and to wipe the user data if needed.

The required monitor dpi is computed using java.awt.Toolkit, but (at least
on MacOS) it doesn't return the correct value, so it can be user supplied.
It's stored at least while the app is running and if possible in the settings
of the ADV/SDK Manager/Updater.

The wipe-data and scale flags are stored and reused while the app keeps
running as well.

Change-Id: Ia2f3ff5f4de285a3d505c6914d6b89cc663be284