Fix Notepadv3Solution orientation changes during NoteEdit
The current NoteEdit.java in Notepadv3Solution crashes on NPE when the
screen orientation changes.  This happens because a null Long is
auto-unboxed to a long when used as an input to
Bundle.putLong(String, long).  The easiest solution is to use
Bundle.putSerializable() instead.

In addition duplicate notepad entries could result because mRowId was not
necessarily defined when onSaveInstanceState(Bundle) was called.  The
solution to that is to call saveState() in that method.

Fixes these buganizer bugs:

Change-Id: Ice325f3b089867e4716deb48aefe8ec03f30ad55
http://b/issue?id=2266994
http://b/issue?id=2266962