Fix make_windows_sdk.
SDK Bug 2478848: emulator fails to build because sdl-config
was not checkout as +x by git/cygwin.

SDK Bug 2479321: aapt built using Cygwin 1.7.x depends on some
new cygwin DLL instead of mgwz.dll. The latter DLL is not even
present and fails to be bundled. This does not change the behavior,
it just makes sure we are using the proper version of cygwin.