Use __asm__ instead of asm in public libc headers
If compiling userland code with -std=c99, the current header produces an
error. The content of this header originally is a kernel internal header,
where asm() is acceptable. In a header visible to userland, this should be
__asm__ instead.

This change is done for all newly extracted headers by the extraction
script, by https://review.source.android.com/19406, which was recently
merged.

Change-Id: I69332124a56577673a14e84cd7d9a4462c307724