Adjust layouts of transition demos to improve fade-outs.
When a container's child count goes to zero, it may report a measurement size
of zero. This affects LayoutTransition's fading-out of target views
because they may get clipped out by a layout that got sized to 0 because it
thinks it has no children. The workaround is to tell the layout to size itself
either with some minimal size (in the case of the FixedGridLayout used by some
demos) or with match_parent in some cases to ensure that the layout always has
some nonzero size.

Change-Id: I908e64d4fb054928af277a021a328e94477c9c83