Android emulator opengles - unit test gles wrapper
This adds opengl es implementation driver that selects whether to use
the native android emulator opengl impelementation or use a remote,
host side renderer. This driver is used to selectivley run applications using
the remote renderer. The selection is based on the application/process name and
is configured using /etc/gles_emul.cfg in the emulator side.

This change also installs /etc/egl.cfg that is used by the Android opengl wrapper
to select an opengl implementation

Change-Id: I66eda3d48938f8057bf45e6476b7faefaf1130fb