Adds the Random Music Player sample.
This sample illustrates how to write an application that plays media
in the background, handles streaming media from the network and
reacts appropriately to system events like headphone disconnection and
audio focus changes, and also uses the notification system to
keep the user informed about what is happening.

This CL is a cherrypick of the old CL 104495 (originally in
honeycomb-mr1) into honeycomb-mr2. Original CL:
https://android-git.corp.google.com/g/#change,104495

Change-Id: Ib8b15ee920300adc093a8b52c17bd9c251ac74ca