sdk: Ensure we copy build.prop to platform image directory.
This patch ensures that we copy the generated build.prop file to
the platform image directory. This file will later be used by the
emulator and other tools to determine various properties of the
image without having to parse the system.img file.

For example, this can be used to extract the target CPU architecture
and/or ABI, and automatically adjust emulation based on this.

Change-Id: Ia5b67c6fb93078b6b224b10e3322af88b99c50b9