Make sure RGB_565 config is available to emulator
There are Android applications (notably BootAnimation) which require
exactly an RGB_565 configuration, and are not satisfied by an e.g.
RGBA_8888 configuration. This doens't make too much sense, and I think
such apps should be fixed.
However, until that happens, there is an issue when the host does not
have an RGB_565 FBConfig, which is typical to many OpenGL implementations.
In such cases, BootAnimation doesn't run.
The workaround is to add an RGB_565 config to the list maintained by the
translator, which is basically an 888 config in disguise.