opengl translator: add native config id to EglConfig
The EGL_CONFIG_ID attribute should be uniqueue for each
config, at some cases we duplicate a configuration in
order to simulate different pixel format (RGB_565) at that
case we want to keep the native config_id of the backend
OpenGL the same on both configs.
This change adds extra attribute 'native_config_id' for that
purpose.

Change-Id: I7e560a4badd01d146b9753cee603ff4707ccfd7f