emulator:opengl: input event redirection
This patch allows the emulator_test_renderer program to redirect
input events (mouse, keyboard) to the emulator's console program.

This in order to make testing easier, while we wait for integration
of the display into the emulator program.

For now, this is hard-coded to use localhost port 5554, and only
a few keys are properly translated (Dpad, HOME, BACK, MENU).

Change-Id: Ie6e37c85c291f53de49220a181cb8724f66519ea