opengles emulator: fix shader parser precision qualifiers
Workaround to intel/linux driver issue.

Desktop OpenGL does not support precision qulifiers in shaders,
so we remove them by defining "precision" "mediump" "lowp" and
"highp" to blank.
That translates the line "precision mediump float;" to "float;" which
is a valid command according to the GLSL spec however it fails
to compile using the intel/linux driver.

This change adds a parser which removes completely the
precision command from the shader source.

Change-Id: I29a16bb78ac0000d624aa2f83e93ef5c6ad60345