CherryPick 676e13 from master to mr0. do not merge.
Move the renderscript include files into platform-tools

The new unique compiler working for all target platforms requires to use
only the latest includes instead of using a different set per target
platforms.

By putting the include in the platform-tools, we ensure that they are
available no matter which platform has been donwloaded and are in sync
with the compiler.

Change-Id: Ib3e8e4e0b02956f60e7f1b4a2e8584a83a33d013