Move the renderscript include files into platform-tools
The new unique compiler working for all target platforms requires to use
only the latest includes instead of using a different set per target
platforms.

By putting the include in the platform-tools, we ensure that they are
available no matter which platform has been donwloaded and are in sync
with the compiler.

Change-Id: Ia5d3d155eeb45c546656123f8cc75d8efe6e631c