SDK: fix script creating platform source archive.
It turns out that the way I use Python to create ZIP
archives is plain wrong. Some files simply don't have the
right size and we get EOFException when unzipping with
Java's ZipFile.

This change adds a flag to instead simply copy all the
files we want and then run the 'zip' system command.
The python zip facility is left there intact in case
I want to go back and fix it later (I'm going to assume
it's my usage that's wrong before really blaming python).

Change-Id: Iea178a49be0bf23c91c01a2e036ae7a76def2b55