ndk: Remove extra GLES symbols on 64-bit platforms
The original arch-arm libGLESv1_CM.so and libGLESv2.so symbols in
android-4 and android-5 were correct. They didn't include some vendor
extension symbols that are in the platform library but that apps
should query at runtime rather than link against.

When arch-x86 and arch-mips were added in android-9, they mistakenly
included these extension symbols. Oh well, for compatibility we can't
fix that now.

The android-20 symbols for all architectures were copied from x86 or
mips, and include the extension symbols. To avoid spreading the
mistake, this change:

(a) Removes the symbol files for architectures that existed in earlier
versions of the platform, since they shouldn't change in android-20.

(b) Removes the extension symbols on 64-bit platforms, so they match
the symbols in arch-arm.

Change-Id: Ib227d9d163bfe2859b13eb57dc19518d3a3f90d0