Add customtabs support lib to extras/android/support/
Currently the only way to get access to customtabs library through
download m2repository and using the aar file whereas all other
support libs are also present in extras/android/support/ with
the corresponding resources and jar file.

This adds the jar and manifest to the corresponding folder to avoid
the confusion to developers.

Change-Id: Id6d71e4071c0725257cadb539cd47230d6727914