Include pre-generated copy of api-versions DO NOT MERGE
This CL changes the api-versions.xml file released in platform tools
to correspond to what was checked into prebuilts rather than what
is being generated by the build. The version generated by the AOSP
build is incorrect for release, but we are now moving to releasing
platform tools from AOSP. In addition, this file will be removed
from platform tools after a few releases (it will be available from
the platform component instead).

Test: make win_sdk sdk_repo and check the included api-versions.xml
Change-Id: I0c3dd3c99ec0c852e7cf8d86b83e298e157044d2