Add an empty directory so we generate android-26.
Without this, we won't copy the android-26 libraries even though we
have them because the platforms directory is built based on what
directories we have here.

It would be better to rewrite gen-platforms.sh to be less awful, but
given that the whole script can be deleted when we drop the
deprecated headers in r16, there's not much point in doing so.

Test: ndk/checkbuild.py && ndk/run_tests.py --filter aaudio
Bug: None
Change-Id: I3c1861f50c8e779d5489e12887180ef1f2099e4a