repo_diff: sync tip of branches of no tag is given
Not all Android source trees support build IDs and tags.
When no build tag is requested, rather than try to match
the build IDs just sync the branch tips of the upstream
and downstream source trees.

Also select two Android Oreo release candidates
as the default branches. The fact that there are frozen
branches and that the diff is small serve as a good showcase
of the tool.

Change-Id: Ifdefd79fb6b32c6b83e0aee390dd5a435b6a61fc