Use mangled name as unique id for C structs as well.
Some structs could be included without extern "C" guards. If the same
struct is included in a C++ source file and a C source file, it's unique
type name will be mangled in  one and its name in the other. To work
around this inconsistency, mangle their names regardless of context.

Bug: 69568963

Test: tests/test.py

Change-Id: Ie285c4691dbaa58389a041fee5116ccbfc7985ce