vndk-def: Refine ineligible dep error message
This commit refines the error messages for ineligible dependencies.  If
a vendor module depends on a vendor_available system shared library but
that system shared library is not copied to /vendor/lib[64], print an
error message to remind developers to copy them.

Bug: 74319871
Test: Manually remove libminijail or libdrm from /vendor/lib[64] and
run `./vndk_definition_tool.py check-dep --system ... --vendor ...`

Change-Id: I6ef2ebe3be137ae0b05f2ba4a8218c72f1ca9ef2