sourcedr: Check whether paths map to projects
This commit checks whether the path maps to a git project specified in
`manifest.xml`.  If it does not map to a git project, emit an error and
skip it.  If we don't do so, an exception will be raised because the
`sorted()` function cannot compare `None` object with `str` object.

Test: Add a project to external and it does not stop the analyzer.
Change-Id: I386ab7c1f76d5630a2c5f43186f38a791a3bb7b9