Display only elements actually visible on the screen as "Visible"
Use the visibleRegion property to determine if a layer is visible on
the screen or not. Elements may have a size and a crop but be displayed
below another.

While the visible option cannot be used to render the elements due to
transforms, it only has a value if the element is displayed on the
screen.

Note that elements may still not appear on the screen as all their
visible regions are below the screen (right < 0 or bottom < 0)

Test: Open trace in winscope, check if only layers visible on the screen have a visible chip and are displayed.
Change-Id: I75137a5057a5bda96db887b110b053e64b49f885