header-abi-linker: Support solib and version script
This commit allows user to specify both `-so` (for a shared object file)
and `-v` (for a version script).  If both of them are specified, the
generated output will only include the symbols that are marked as
exported in both cases.

This commit also rewritten the version script parser so that it can
work with real version script cases.  This includes how the parser
skips a comment line.

Furthermore, the version script parser accepts a list of excluded symbol
versions and symbol tags that should not be included in the generated
output.  To customize the behavior, pass `--exclude-symbol-version`
and/or `--exclude-symbol-tag` to `header-abi-linker`.

Bug: 122845490
Test: ./tests/test.py
Test: atest  # under development/vndk/tools/header-checker
Change-Id: Ibf048bba6ccd5ac44f265d1fe767c2ded3d13a8d