Add Type to transform objects.
Add a type, based on the framework/native definitions, to the transform
objects.

If the transform matrix can be reconstructed from the type (identity,
translate, rotate 0/90/180/270) it should not be stored on the protobuf,
but should instead be reconstructed on Winscope.

Test: Record a Proto3 SF trace. Open it without displaying default
values in Winscope. Check if it is correct. Select a legacy trace. Open
it in Winscope. Check is it is also correct.

Change-Id: I03744efba2c8740d7c543f6793bce0ba7bf6d975