Added visibility check for missing relativeOf
When the layer has the flag isRelativeOf but the relativeOf is -1, that
means the relativeOf layer has been removed. This would prevent the
layer from showing up on screen, but still be in memory. Added this
reason to list of reasons why layer wouldn't be visible.

Also added fix for BufferLayer check to instead check for BufferQueueLayer
and BufferStateLayer.

Test: winscope shows "RelativeOf layer has been removed"
Change-Id: Id1d59bd71a5aac7a011204ba345d694f67c28a29