Remove old GNU linkers from SDK.
Linkers aarch64-linux-android-ld and x86_64-linux-android-ld were built
with a minimum macOS deployment target lower than 10.9, so they cannot
be signed and notarized.  Rather than try to produce newer versions of
these linkers (which are obsolete in the SDK anyway), we remove them (so
that all of the SDK can be signed and notarized).  Rather than introduce
an inconsistency among the linkers for a particular host (there are
three other GNU linkers that don't have this problem) or across hosts
(there are no issues on linux or windows), we remove all of the GNU
linkers, not just the two problematic ones.

Only newer gradle plugins (those that look for the clang linker lld)
will work with the new SDK from which those old linkers have been
removed.

Bug: 152337684

Test: (gLinux) m TARGET_PRODUCT=sdk_phone_armv7 TARGET_BUILD_VARIANT=userdebug sdk dist sdk_repo
      (gLinux) m TARGET_PRODUCT=sdk_phone_armv7 TARGET_BUILD_VARIANT=userdebug win_sdk dist sdk_repo
      Inspect sdk-repo-{linux,windows}-build-tools-eng.*.zip
Change-Id: Iccfa870a826de3f12c99175e0761ea00fe2876ed