Add fallback mechanism for identifier of older window traces
Older WM traces might not have an identifier set for
WindowContainer since this was added in ag/11659380.

This change makes Winscope again backwards compatible with older
WM traces. If it exists, the identifier of a parent holds the
place of a missing WindowContainer identifier. It is possible
that both identifiers are null, in which case the title and the
hashCode of WindowContainer are set as empty strings.

Bug: 170372278
Test: build Winscope locally, upload a WM trace and visualise it
Change-Id: Ieb5adc1d6a94d9a9215304b92114b807f1d812b4