Include 'name' in props; use post-sanitize module name.
In cases where not all the variants of a library are
included in the generated vendor snapshot, be sure to
include the module name in the props so that it can be
factored out later.

Conversely, be sure to compute the notice filename from
the de-sanitized module name so so that it can also be
factored out later.

Bug: 171413810
Bug: 157967325
Test: Verify that full snapshot Android.bp unchanged.
Test: Verify correct syntax for filtered snapshot.
Change-Id: I5725e17b19907d6a5063823e8b99bd0517c7c144