Make FASS accessible from winscope
Flickerlib has been refactored so that more functionality can be used directly in Winscope.

This will enable us to add an option in winscope to detect transitions
for manually created WM and SF traces, without relying on flicker.

The winscope change is not part of this topic.

This CL updates winscope for compatibility

Test: build and check if winscope works
Change-Id: I28dca32600786b7955afa98f632f06628b528222