Create and show Chrome Wayland logs for Winscope
This change is adding Chrome Wayland logs to Winscope. The logs are
created by Winscope and shown on it. Since Wayland logs have been
already compatible with Winscope, this change adds some configs to
existing ones. Also, "Wayland" option is only visible when we connect
ARC, so Android users don't need to pay attention to conflicts.

Bug: 201374806

Test: manual by follow steps
    1) Start tracing by choosing "Wayland" and clicking "START TRACE" on Winscope.
    2) End tracing by clicking "END TRACE".
    3) See if the insets logs are available.
    4) Check if "Wayland" option is only visible when we connect ARC.

Change-Id: Ibfb5564a32c75929b449915a75db23ae3cd6447e