Stop cargo2android.py from printing the --cargo_bin argument.
cargo2android.py prints out all of its arguments to the top of the
Android.bp file (and into a config file if that option was set) as a
way of persisting them.  However, we do not want to persist the
--cargo_bin directory, as this is a proprety of the local setup and
not the crate itself.  Thus ensure we don't print this.

Test: Run with argument.
Change-Id: If5292d1d740b08755ba6d31654239e93232fdc49