Limit scope of PropertiesTableView styling
Previously in [1], the style in PropertiesTableView component is applied
directly to common Vue components like <md-table-cell />. This may
create confusion in the future if <md-table-cell /> is used in other
components, as the style configuration of <md-table-cell /> would not
be the default Vue one, and would instead follow PropertiesTableView's.
This commit fixes this issue by limiting the scope of the CSS styling
in PropertiesTableView.vue.

CL[1]: I39fe382ffb50d71e67cace912452a33ca49de3f9

Test: manual on local build of Winscope
Bug: 235313782
Change-Id: I44d9afb8f49cc21748807c0b8b7e912db6f37a52