[11] Change matching WM/SF entry to IME entry algo
Due to there being some slightly different timestamps between WM/SF
trace and IME trace, we want to allow some "fault tolerance" -- even
if a WM/SF entry comes after IME entry in time, it may still be the
corresponding entry if it's within a manually-defined "fault tolerance"
time.

For now, fault tolerance is defined to be 200ms. So if a WM/SF entry is
within IME entry +- 200ms, it may be considered a matching /
corresponding entry. Within this range, we will always try to pick a
WM/SF entry that comes before the IME entry. If there is no such WM/SF
entry, then we will pick one that comes after.

Bug: 236226833
Test: manual on local build of Winscope
Change-Id: I0782ec520284517bf964eeffd552e6175758c9f1