Update isLockscreenDismissed when swiping in STL demo
This CL fixes a bug that was introduced in
I92d98a1a5e4abd0a4d8d8506cc5d0e6a4527ad56: now that we are using
MutableSTLState, onChangeScene() is not called when committing scene
transitions. Instead, we have to rely on canChangeScene (when it returns
true) to correctly update isLockscreenDismissed.

Bug: 308961608
Test: Manual, swipe from unlockable lockscreen to launcher, then to
 shade, then to launcher.
Flag: EXEMPT demo app
Change-Id: Iab9db725eebd9814c4d6ac67856bf36d80c287d8