Improve handling of corrupted/empty traces.
Discard empty traces at the parser creation stage. Assert that when a parser is created in TracesParserFactory, AbstractParser (perfetto) or ParserFactory (legacy), there is at least one entry.

Improve user reporting of loaded files that are corrupted/do not have visualisation. Do not allow users to proceed with "View traces" if all loaded traces are corrupted/not visualisable.

Bug: 319302757
Test: npm run test:unit:ci

Change-Id: Idc6fdc9379bc2f6865650575cc3de187b5c5cd29