Add clipping and scroll to shades
This CL adds clipping and scrolling to the QuickSettingsShade and
NotificationShade in the STL demo.

The main difficulty with enabling clipping to a shade that is coming
from the top of the screen is that shared elements that go from their
source source to their destination scene will be clipped as well. To
solve this, this CL relies on a new API implemented in ag/29200580:
sharedElement(..., elevateInContent = /* contentKey */). This allows the
shared elements to escape the clipping applied in their overlays.

Bug: 373799480
Test: Manual, in the STL demo
Flag: EXEMPT demo app
Change-Id: I38c3d7ec0f893bef1431194faac89bbb25eb1fbb