Winscope: Use input frame as input window rect
The input frame is usually, but not always, the same as the layer
bounds. Previously, we were using the layer bounds as the input window
rect, but we should be using the input frame instead. This will help us
debug cases where the input frame isn't populated correctly.

Note that input frame and touchable region are in a different coordinate
space (display space) compared to the layer bounds, so they must be
transformed accordingly.

Bug: 376087877
Test: manual, with traces
Test: presubmit
Flag: EXEMPT dev tool
Change-Id: I2ce0bc99549be5a259429670f1f794deb82c3d04