Add colored border around demo STL
This CL adds a border around the demo STL that is green when we are idle
and that is red when transitioning. This can help visualize when a
transition is done vs still ongoing. It  also helped me observe a bug
with a gesture transition that was never stopped even after removing my
finger with some pending changes for b/379281707.

Bug: 379281707
Test: Manual
Flag: EXEMPT demo app
Change-Id: Icad1365024380a8db7df65761e417b5659321d6a