Use one AudioRecord per VirtualDevice
Instead of relying on 2 AudioRecords, one just for managing the audio
policies and one for the actual audio data transfer, just reuse
the first created AudioRecord (which is still always open/recording).
This reduces mix recreation impact introduced by ag/27671536 under the
com.android.media.audioserver.fix_input_sharing_logic flag.
It also works as expected with and without the flag enabled.

This practically fixes the audio playback restart after pause on/from
a remote virtual device. Using a manual flush on the always running AudioRecord, the stream is now playing on the client device with no glitches when pausing a playback.

Bug: 358151598
Test: From VdmHost app (with audio client setting enabled) create a
virtual device on a client with audio capabilities. Play an audio stream
(e.g. local host media files in the Files app) and pause and unpause the
playback.
Change-Id: Ie4a7da43159be6e70c1b6809df6ef14f38b33765