Winscope: InputViewer: Propagate unpopulated coordinate values
A single input event can be dispatched to multiple windows, where each dispatch
is represented by a separate AndroidWindowInputDispatchEvent. The dispatch event
can contain axis values in the window and raw coordinate spaces, but to save
space, these values are only traced if they are different from the events'. In
this operation, we propagate the coordinate values in the cases that they were not
logged to save space.

Bug: 388336752
Test: Presbumit
Flag: EXEMPT dev tool only
Change-Id: If58d408cc24e7692a0fbe35295fd71f81f4c017f