Read from the AudioRecorder(s) in RecorderDemoActivity
In order to activate the dynamic audio policies used for VDM, the proper
routing is done after data is being read from the AudioRecord(s).
So wrapped the AudioRecord(s) in threads and actual read data from them.
The audio data is not used for now.

Bug: 381386115
Flag: EXEMPT sample app dev
Test: manual VdmDemos app usage
Change-Id: I8228276468d0fe078d228526b6d4d4b55163e84c