Add extra to change the number of notifications in STL demo
This CL adds an extra to make it possible for benchmark tests to change
the number of notifications inside the STL demo. This will be used to
increase the number of STLs in the start-up test, allowing us to better
measure the intial cost of composing STL (given that each notification
is its own STL).

Bug: 389985793
Test: Manual
Flag: EXEMPT demo app
Change-Id: I3eab911fd520878520b146dce21319727958c97b