When updating crates, update dependent crates as well.
This is controlled by 'update_with' in android_config.toml.

Also get rid of the DENYLIST for the crate updater, since most of them
are now handled by this new feature of crate_tool. For the remainder, we
can rely on the cache of recent updates to avoid retrying them too
often.

Test: ./crate_tool update argh 0.1.13
Change-Id: Ibf5af678b2c264a46be5e1b48f4948261d3bdd9c